browser-agent/python/requirements.txt
# LangChain Core
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0

# MCP Protocol
mcp>=1.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=12.0

# HTTP Client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
anyio>=4.0.0
